Add files via upload
Trained the sequence to sequence model (Model 1) for a language pair1 where the output is English and the
input is Hindi. Then trained another model (Model 2) for the reverse (i.e., from English to Hindi). In this model, I have used the GloVe 100-dimensional embeddings
while training. Used the translated English sentences to again recreate the original Hindi sentence.